feat: integrate typeshare with a custom build.rs
We wanted to generate use tauri-specta to generate TypeScript types
but some special types like chrono::DateTime we cannot derive the Type
trait because that's an external type. I searched another solution
and typeshare seems to do the work for types, it generates types
even though subtypes have not been annoted. There is no library, so I had
to run the CLI and add it as dependency.