feat(desktop): start to build a TypeScript LiveClient
I was using the Rust LiveClient in Tauri commands but when I started implementing
streaming of a subset of events, it was just impossible to find a good pattern
to have streaming of these events, in addition to continue to send
Action on the socket. Managing states in Tauri and sending back to the frontend
was also a pain to manage.