feat: integrate the PLX DY spec into the FromDir course and exo implementations
This integrates the DY library + the PLX DY spec developed in an external Git
repository. I removed the FromDir implementation for Skill as it can no longer
be loaded from a folder as there is a list of skills under skills.dy, instead
of the individual skill.toml. For now, there is no way to control the order
of exos inside skill because of this change. This is made in the style of Delibay
course structure.