fix: cancel previous workers when saving file
First this avoids mixing different results in the interface, results from different executions
being sent at different times.

This also fixes the issue where the compilation fails on Windows because
the executable file cannot be written because the existing file is running.
This was easily reproducible by adding sleep() in the program to have a slow execution
and a faster build when saving the second time. I